requirements.txt
gradio>=4.0.0
opencv-python>=4.8.0
pillow>=10.0.0
openai-whisper>=20231117
transformers>=4.30.0
torch>=2.0.0
googletrans==4.0.0rc1 